worker-backend/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Flores",email = "dnielfs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector] (>=1.0rc,<2.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
